transferred Radio to its own component
