+++ client/components/Radio/radio.scss
@@ -1,0 +1,11 @@
+.radio {
+  height: 16px;
+  width: 16px;
+  margin-right: 6px;
+  fill: #878a8c;
+
+  &:active,
+  &--checked {
+    fill: $blue;
+  }
+}
