+++ client/pages/subreddits/create/create.scss
@@ -107,17 +107,6 @@
         margin-bottom: 16px;
         cursor: pointer;
 
-        &__radio {
-          height: 16px;
-          width: 16px;
-          margin-right: 6px;
-          fill: #878a8c;
-
-          &:active {
-            fill: #0079d3;
-          }
-        }
-
         &__icon {
           height: 16px;
           width: 16px;
